% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_tutorial.R
\name{add_tutorial}
\alias{add_tutorial}
\title{Add a tutorial to a project}
\usage{
add_tutorial(name, package)
}
\arguments{
\item{name}{A character string. The name of a tutorial saved in a package.}

\item{package}{A character string. The name of the package that contains a
tutorial.}
}
\description{
Warning: \code{add_tutorial} does not work as written because it assumes that
R will recognize that RStudio IDE is running when R evaluates .Rprofile. This
is an incorrect assumption. However, \code{add_tutorial} and
\code{remove_tutorial} should be easy to adapt to a an RStudio-specific
startup file should RStudio implement one.
}
\details{
\code{add_tutorial()} inserts a call to \code{\link{view_tutorial}()}
into the .Rprofile file contained in the current working directory. As a
result, R will launch the tutorial in the RStudio IDE viewer pane whenever
the current project is opened or re-opened.

If an .Rprofile file does not exist in the current working directory,
\code{add_tutorial()} will create one. If the .Rprofile file in the current
working directory already contains a call to \code{view_tutorial()},
\code{add_tutorial()} will return an informative error message. Note that a
user may also have .Rprofile files in both their home directory and their R
home directory. If one of these files contains a call to
\code{view_tutorial()}, \code{add_tutorial} will override the call without a
warning message.
}
\seealso{
\code{\link{view_tutorial}}
}
