% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_code.R
\name{check_code}
\alias{check_code}
\title{Check code structure}
\usage{
check_code(correct = "{random_praise()} Correct!",
  incorrect = "{message} {random_encourage()}", grader_args = list(),
  learnr_args = list())
}
\arguments{
\item{correct}{A character string to display if the student answer matches
the solution code.
This character string will be run through \code{glue::\link[glue]{glue_data}} with \code{list(correct = TRUE, message = NULL)}.}

\item{incorrect}{A character string to display if the student answer matches
the solution code.
This character string will be run through \code{glue::\link[glue]{glue_data}} with \code{list(correct = FALSE, message = "<STRING>")} where message is the error found while comparing the user solution to the known solution.}

\item{grader_args}{A list of parameters passed to \code{grader} functions (provided by \code{grade_learnr}).
This can be student code to check against the \code{results} surrounded
by \code{quote()}, \code{rlang::quo()}, or provided as a character string.
For example, \code{user_quo = quote(1)}}

\item{learnr_args}{A list of parameters passed to \code{learnr} (provided by \code{grade_learnr}), for example, \code{last_value = 1}}
}
\value{
a \code{grader_graded} structure from \code{\link{result}}.
  If the student answer differs from the
  solution code, the message will describe the first way that the answer
  differs, and it will ask the student to try again. If the answer matches
  the solution code, the message will be the content of the \code{correct}
  argument.
}
\description{
\code{check_code()} compares user code to a solution (i.e. model code) and
describes the first way that the user code differs. If the user code exactly
matches the solution, \code{check_code()} returns a customizable success
message.
}
\details{
\code{check_code()} provides a *strict* check in that the user code must
exactly match the solution. It is not enough for the user code to be
equivalent to the solution code (e.g. to return the same result as the
solution).

You can provide solution code for \code{check_code()} to use in two ways:

1. Pass code as a character string or a quoted expression to the solution
argument of \code{check_code()}

2. Make a "-solution" code chunk for the exercise to be checked in a learnr
document. There is no need to supply a solution argument for
\code{check_code()} if you call it from the "-check" chunk of the same
exercise. Likewise, there is no need to supply a user argument when you call
\code{check_code()} from a learnr document (learnr will provide the code
that the student submits when it runs \code{check_code()}.

For best results, name all arguments provided in the solution code.
}
\examples{
\dontrun{grading_demo()}
}
