% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_result.R
\name{check_result}
\alias{check_result}
\title{Check result of exercise code}
\usage{
check_result(..., correct = NULL, incorrect = NULL,
  grader_args = list(), learnr_args = list(),
  glue_correct = getOption("gradethis_glue_correct"),
  glue_incorrect = getOption("gradethis_glue_incorrect"))
}
\arguments{
\item{...}{\code{\link{pass_if()}} or \code{\link{fail_if()}} \code{\link{condition}}s to check}
}
\value{
a \code{\link{graded}} object from either
  \code{\link{pass_if}} or \code{\link{fail_if}} containing a formatted
  \code{correct} or \code{incorrect} message and whether or not a match was found.
}
\description{
\code{check_result()} compares the final result of the student code to known
\code{\link{pass_if()}} and \code{\link{fail_if()}} \code{\link{condition}}s.
If the student result exactly matches a known case, \code{check_result}
returns the matching message value.
}
\examples{
\dontrun{grading_demo()}

check_result(
  pass_if(~ .result == 5, "This is a correct message"),
          grader_args = list(),
          learnr_args = list(last_value = 5, envir_prep = new.env())
)

check_result(
  pass_if(function(x) x == 5, "This is a correct message"),
  learnr_args = list(last_value = 5)
)

check_result(
  pass_if(5, "This is a correct message"),
          learnr_args = list(last_value = 5)
)

check_result(
  fail_if(5, "You were supposed to do this other thing!"),
          pass_if(~ TRUE, "should never reach here"),
          learnr_args = list(last_value = 5)
)
}
\seealso{
\code{\link{check_code}}, \code{\link{check_result}}, and \code{\link{test_result}}
}
