% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{condition}
\alias{condition}
\title{Condition object}
\usage{
condition(x, message, correct)
}
\arguments{
\item{x}{expression to be evaluated}

\item{message}{character string for message returned (usually passed in from
\code{\link{pass_if}} or \code{\link{fail_if}}}

\item{correct}{logical whether the condition is the correct answer}
}
\value{
a \code{grader_condition} object that contains
  the expression \code{x},
  the message \code{message},
  whether or not the expression is the correct answer or not, \code{correct},
  the type of expression (formula, function, or value), \code{type}
}
\description{
Captures what the student passes into \code{\link{pass_if}} or \code{\link{fail_if}},
figures out what type of object was passed into \code{x},
and returns a \code{grader_condition} object that will be passed into
\code{\link{evaluate_condition}}.
}
\examples{
\dontrun{
  condition(~ identical(x = .result, 5), message = "Correct", correct = TRUE)
}
}
