% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{correct}
\alias{correct}
\alias{incorrect}
\alias{is_rrect}
\title{Correct or Incorrect result}
\usage{
correct(message = NULL)

incorrect(message = NULL)

is_rrect(x)
}
\description{
Correct or Incorrect result
}
