% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_mistakes_with_infix.R
\name{detect_mistakes_with_infix}
\alias{detect_mistakes_with_infix}
\title{detect_mistakes_with_infix}
\usage{
detect_mistakes_with_infix(user, solution, .name = NULL)
}
\arguments{
\item{user}{}

\item{solution}{}

\item{.name}{}
}
\description{
detect_mistakes_with_infix
}
