% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_learnr.R
\name{enable_gradethis}
\alias{enable_gradethis}
\title{Enable gradethis in learnr}
\usage{
enable_gradethis()
}
\description{
Call this function inside a learnr document to use gradethis' suggested
setup for grading exercises.
}
\seealso{
\code{\link[=grade_learnr]{grade_learnr()}}
}
