% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graded_condition.R
\name{eval_gradethis}
\alias{eval_gradethis}
\title{Capture grades and errors}
\usage{
eval_gradethis(expr, on_error = NULL, on_graded = NULL)
}
\arguments{
\item{expr}{The expression or code block to evaluate}

\item{on_error}{A \code{\link[=withCallingHandlers]{withCallingHandlers()}} handler for class \code{error} with
signature \verb{function(error, this_env)} that receives the error object and
calling environment of the error handler. \code{on_error} should use
\code{\link[rlang:return_from]{rlang::return_from()}} using \code{this_env} to immediately return the value and
not continue evaluation.}

\item{on_graded}{A \code{\link[=withCallingHandlers]{withCallingHandlers()}} handler for class \code{graded} with
signature \verb{function(grade, this_env)} that receives the error object and
calling environment of the error handler. \code{on_graded} should use
\code{\link[rlang:return_from]{rlang::return_from()}} using \code{this_env} to immediately return the value and
not continue evaluation.}
}
\description{
Capture the first \code{\link[=graded]{graded()}} signal or error thrown when evaluating the
\code{expr}.
}
\examples{
# Passes with "message 1", short-circuiting evaluation
eval_gradethis({
  pass("message 1")
  pass("message 2")
  pass("message 3")
})

# Fails with message from fail()
eval_gradethis({
  fail("incorrect")
  pass("correct")
})

# Fails with message from expect_true()
eval_gradethis({
  testthat::expect_true(FALSE)
  pass("message 2")
  pass("message 3")
})

# Fails immediately with message "boom"
eval_gradethis({
  stop("boom")
  pass("message 2")
  pass("message 3")
})
  
}
\keyword{internal}
