% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_condition.R
\name{evaluate_condition}
\alias{evaluate_condition}
\title{Evaluates a condition}
\usage{
evaluate_condition(condition, ..., last_value, env)
}
\arguments{
\item{condition}{a \code{\link[=condition]{condition()}} object}

\item{...}{ignored}

\item{last_value}{The last value from evaluating the user's exercise submission.}

\item{env}{environment to evaluate the condition}
}
\value{
a \code{\link[=graded]{graded()}} value if \code{condi$x} is \code{TRUE} or
\code{NULL} if \code{condi$x} is \code{FALSE}
}
\description{
Evaluates the \code{\link[=condition]{condition()}} object to return a \code{\link[=graded]{graded()}} value.
}
\examples{
condi_formula_t <- condition(
  ~ identical(.result, 5),
  message = "my correct message",
  correct = TRUE
)
evaluate_condition(
  condi_formula_t,
  last_value = 5,
  env = new.env()
)
}
