% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass_fail_if.R
\name{fail_if}
\alias{fail_if}
\title{Fail if condition matches}
\usage{
fail_if(x, message = NULL)
}
\arguments{
\item{x}{A formula, function, or value, that returns \code{TRUE} or \code{FALSE}.
When comparing objects that are greater than length 1 (e.g., vectors, dataframes, matricies, etc)
A boolean vector will be returned if the user uses \code{==}, not a single boolean value.
\code{grader} will run the vector through \code{all(..., na.rm = TRUE)} to check for the boolean value.
It is advised that the user use \code{identical()} instead of \code{==} in this case.}

\item{message}{chracter string for message returned}
}
\description{
Fail if condition matches
}
