% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_result.R
\name{find_first_condition}
\alias{find_first_condition}
\title{Find a matching condition}
\usage{
find_first_condition(..., grader_args, learnr_args)
}
\arguments{
\item{...}{\code{\link[=pass_if]{pass_if()}} or \code{\link[=fail_if]{fail_if()}} \code{\link[=condition]{condition()}}s}

\item{grader_args}{A list of parameters passed to \code{grader} functions
(provided by \code{\link[=grade_learnr]{grade_learnr()}}). This contains:
\itemize{
\item \code{user_quo}: Quoted R code submitted by the user. For example
\code{\link[rlang:quo]{rlang::quo(1)}}
\item \code{solution_quo}: (Optional) Quoted solution R code provided by the
\verb{*-solution} chunk for an exercise.
}}

\item{learnr_args}{A list of all parameters passed to \code{\link[=grade_learnr]{grade_learnr()}} by
\code{learnr}. See
\url{https://rstudio.github.io/learnr/exercises.html#exercise_checking} for
more details.}
}
\value{
The first condition that matches. If no conditional matches, returns
\code{NULL}.
}
\description{
Looks through a given set of conditions to attempt to find a match.
}
