% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_result.R
\name{find_matching_condition}
\alias{find_matching_condition}
\title{Find a matching condition}
\usage{
find_matching_condition(results, grader_args, learnr_args)
}
\arguments{
\item{results}{A list of \code{\link[=pass_if]{pass_if()}} or \code{\link[=fail_if]{fail_if()}} \code{\link[=condition]{condition()}}s}

\item{grader_args}{A list of parameters passed to \code{grader} functions
(provided by \code{\link[=grade_learnr]{grade_learnr()}}). This contains:
\itemize{
\item \code{user_quo}: Quoted R code submitted by the user. For example
\code{\link[rlang:quo]{rlang::quo(1)}}
\item \code{solution_quo}: (Optional) Quoted solution R code provided by the
\verb{*-solution} chunk for an exercise.
}}

\item{learnr_args}{A list of all parameters passed to \code{\link[=grade_learnr]{grade_learnr()}} by
\code{learnr}. See
\url{https://rstudio.github.io/learnr/exercises.html#exercise_checking} for
more details.}
}
\value{
A list with two elements:
\itemize{
\item \code{matching_conditional} - \code{TRUE} if the given list of conditions contained a match. Otherwise, FALSE
\item \code{result} - The evaluated \code{\link[=condition]{condition()}} object; a \code{\link[=graded]{graded()}} value.
}
}
\description{
Looks through a given set of conditions to attempt to find a match.
}
