% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_solutions.R
\name{format.solutions_test}
\alias{format.solutions_test}
\title{Format solutions test output}
\usage{
\method{format}{solutions_test}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\description{
Format solutions test output
}
