% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graded.R
\name{grade_feedback}
\alias{grade_feedback}
\title{Provide grade feedback}
\usage{
grade_feedback(
  grade,
  type = c("auto", "success", "info", "warning", "error", "custom"),
  location = c("append", "prepend", "replace")
)
}
\arguments{
\item{grade}{a \code{\link[=graded]{graded()}} object.}

\item{type}{Feedback type (visual presentation style). Can be "auto", "success", "info", "warning", "error", or "custom".
Note that "custom" implies that the "message" field is custom HTML rather than a character vector.}

\item{location}{Location for feedback ("append", "prepend", or "replace").}
}
\description{
Creates a feedback object suitable for returning in a learnr checking function
(e.g., the \code{exercise.checker} option in \code{\link[learnr:tutorial_options]{learnr::tutorial_options()}})
}
