% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_this.R
\name{grade_this-objects}
\alias{grade_this-objects}
\alias{.result}
\alias{.user}
\alias{.last_value}
\alias{.solution}
\alias{.solution_all}
\alias{.user_code}
\alias{.solution_code}
\alias{.solution_code_all}
\alias{.envir_prep}
\alias{.envir_result}
\alias{.envir_solution}
\alias{.evaluate_result}
\alias{.label}
\alias{.stage}
\alias{.engine}
\title{Checking environment objects for use in \code{grade_this()}}
\usage{
.result

.user

.last_value

.solution

.solution_all

.user_code

.solution_code

.solution_code_all

.envir_prep

.envir_result

.envir_solution

.evaluate_result

.label

.stage

.engine
}
\description{
\code{\link[=grade_this]{grade_this()}} allows instructors to determine a grade and to create custom
feedback messages using custom R code. To facilitate evaluating the
exercise, \code{\link[=grade_this]{grade_this()}} makes available a number of objects that can be
referenced within the \code{{ ... }} expression.

All of the objects provided by \code{learnr} to an exercise checking function
are available for inspection. To avoid name collisions with user or
instructor code, the names of these objects all start with \code{.}.
\itemize{
\item \code{.label}: The exercise label.
\item \code{.engine}: The exercise engine, typically 'r'.
\item \code{.last_value}: The last value returned from evaluating the user's exercise submission.
\item \code{.solution_code}: A string containing the code provided within the \verb{*-solution} chunk for the exercise.
\item \code{.user_code}: A string containing the code submitted by the user.
\item \code{.check_code}: A string containing the code provided within the \verb{*-check} or \verb{*-code-check} chunk for the exercise.
\item \code{.envir_prep}: A copy of the R environment after running the exercise setup code and before the execution of the student's submitted code.
\item \code{.envir_result}: The R environment after running the student's submitted code.
\item \code{.envir_solution}: The R environment after running the solution code.
\item \code{.evaluate_result}: The return value from the \code{\link[evaluate:evaluate]{evaluate::evaluate()}} function (see learnr's documentation).
\item \code{.stage}: The current checking stage in the learnr exercise evaluation lifecycle: 'code_check', 'error_check', or 'check'
}

In addition, \pkg{gradethis} has provided some extra objects:
\itemize{
\item \code{.user}, \code{.result}: The last value returned from evaluating the user's exercise submission.
\item \code{.solution}: The last value returned from evaluating the \code{.solution_code} for the exercise (evaluated in \code{.envir_prep}).
\item \code{.solution_all}: A list containing all solutions when multiple solutions are provided in the \verb{*-solution} chunk for the exercise. Solutions are separated by header comments, e.g. \verb{# base_r ----}.
\item \code{.solution_code_all}: A list containing the code of all solutions when multiple solutions are provided in the \verb{*-solution} chunk for the exercise. Solutions are separated by header comments, e.g. \verb{# base_r ----}.
}
}
