% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_this_code.R
\name{grade_this_code}
\alias{grade_this_code}
\title{Grade student code against a solution}
\usage{
grade_this_code(
  correct = getOption("gradethis.code_correct", getOption("gradethis.pass",
    "Correct!")),
  incorrect = getOption("gradethis.code_incorrect", getOption("gradethis.fail",
    "Incorrect")),
  ...,
  allow_partial_matching = getOption("gradethis.allow_partial_matching", TRUE)
)
}
\arguments{
\item{correct}{A \code{glue}-able character string to display if the student
answer matches a known correct answer.}

\item{incorrect}{A \code{glue}-able character string to display if the student
answer does not match the known correct answer. Use \code{code_feedback()} in
this string to control the placement of the auto-generated feedback message
produced by comparing the student's submission with the solution.}

\item{...}{Ignored}

\item{allow_partial_matching}{A logical. If \code{FALSE}, the partial matching of
argument names is not allowed and e.g. \code{runif(1, mi = 0)} will return a
message indicating that the full formal name \code{min} should be used. The
default is set via the \code{gradethis.allow_partial_matching} option, or by
\code{\link[=gradethis_setup]{gradethis_setup()}}.}
}
\value{
Returns a function whose first parameter will be an environment
containing objects specific to the exercise and submission (see \strong{Available
variables}). For local testing, you can create a version of the expected
environment for a mock exercise submission with \code{\link[=mock_this_exercise]{mock_this_exercise()}}.
Calling the returned function on the exercise-checking environment will
evaluate the grade-checking \code{expr} and return a final grade via \code{\link[=graded]{graded()}}.
}
\description{
\code{grade_this_code()} compares student code to a solution (i.e. model code) and
describes the first way in which the student code differs. If the student
code exactly matches the solution, \code{grade_this_code()} returns a customizable
success message (\code{correct}). If the student code does not match the solution,
a customizable incorrect message (\code{incorrect}) can also be provided.

In most cases, to use \code{grade_this_code()}, ensure that your exercise has a
\code{-solution} chunk:\preformatted{```\{r example-solution\}
sqrt(log(1))
```
}

Then, call \code{grade_this_code()} in your exercise's \code{-check} chunk:\preformatted{```\{r example-check\}
grade_this_code()
```
}

Learn more about how to use \code{grade_this_code()} in the \strong{Details} section
}
\section{Details}{


\code{grade_this_code()} only inspects for code differences between the
student's code and the solution code. The final result of the student code
and solution code is ignored. See the \strong{Code differences} section of
\code{\link[=code_feedback]{code_feedback()}} for implementation details on how code is determined to
be different.

You can call \code{grade_this_code()} in two ways:
\enumerate{
\item If you want to check the student's code without evaluating it, call
\code{grade_this_code()} in the \verb{*-code-check} chunk.
\item To return grading feedback in along with the resulting output of the
student's code, call \code{grade_this_code()} in the \verb{*-check} chunk of the
exercise.
}

To provide the solution code, include a \verb{*-solution} code chunk in the
learnr document for the exercise to be checked. When used in this way,
\code{grade_this_code()} will automatically find and use the student's
submitted code — \code{.user_code} in \code{\link[=grade_this]{grade_this()}} — as well as the solution
code — \code{.solution_code} in \code{\link[=grade_this]{grade_this()}}.
}

\section{Custom messages}{


You can customize the \code{correct} and \code{incorrect} messages shown to the user
by \code{grade_this_code()}. Both arguments accept template strings that are
processed by \code{\link[glue:glue]{glue::glue()}}. If you provide a custom template string, it
completely overwrites the default string, but you can include the
components used by the default message by adding them to your custom
message.

There are four helper functions used in the default messages that you may
want to include in your custom messages. To use the output of any of the
following, include them inside braces in the template string. For example
use \code{{code_feedback()}} to add the code feedback to your custom \code{incorrect}
message.
\enumerate{
\item \code{\link[=code_feedback]{code_feedback()}}: Adds feedback about the first observed difference
between the student's submitted code and the model solution code.
\item \code{\link[=pipe_warning]{pipe_warning()}}: Informs the user that their code was unpiped prior to
comparison. This message is included by default to help clarify cases
when the code feedback makes more sense in the unpiped context.
\item \code{\link[=random_praise]{random_praise()}} and \code{\link[=random_encouragement]{random_encouragement()}}: These praising and
encouraging messages are included by default in correct and incorrect
grades, by default.
}
}

\examples{
# For an interactive example run: gradethis_demo()

# # These are manual examples, see grading demo for `learnr` tutorial usage

grade_this_code()(
  mock_this_exercise(
    .user_code     = "sqrt(log(2))", # user submitted code
    .solution_code = "sqrt(log(1))"  # from -solution chunk
  )
)

grade_this_code()(
  mock_this_exercise(
    # user submitted code
    .user_code     = "runif(1, 0, 10)",
    # from -solution chunk
    .solution_code = "runif(n = 1, min = 0, max = 1)"
  )
)

# By default, grade_this_code() informs the user that piped code is unpiped
# when comparing to the solution
grade_this_code()(
  mock_this_exercise(
    # user submitted code
    .user_code     = "storms \%>\% select(year, month, hour)",
    # from -solution chunk
    .solution_code = "storms \%>\% select(year, month, day)"
  )
)

# By setting `correct` or `incorrect` you can change the default message
grade_this_code(
  correct = "Good work!",
  incorrect = "Not quite. {code_feedback()} {random_encouragement()}"
)(
  mock_this_exercise(
    # user submitted code
    .user_code     = "storms \%>\% select(year, month, hour)",
    # from -solution chunk
    .solution_code = "storms \%>\% select(year, month, day)"
  )
)
}
\seealso{
\code{\link[=code_feedback]{code_feedback()}}, \code{\link[=grade_this]{grade_this()}}, \code{\link[=mock_this_exercise]{mock_this_exercise()}}
}
