% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{graded}
\alias{graded}
\title{Graded submission value
\code{grader_graded} object}
\usage{
graded(correct, message = NULL)
}
\arguments{
\item{correct}{A boolean value of whether or not the checked code is correct.}

\item{message}{A character string of the message to be displayed.}
}
\description{
returned by every \code{check_*} function.
When implementing checking functions to be used by \code{\link{grade_learnr}}, the method should return a \code{graded} value
}
