% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradethis-package.R
\docType{package}
\name{gradethis-package}
\alias{gradethis}
\alias{gradethis-package}
\title{gradethis: Automated Feedback for Student Exercises in 'learnr'
    Tutorials}
\description{
Pairing with the 'learnr' R package, 'gradethis'
    provides multiple methods to grade 'learnr' exercises.  To learn more
    about 'learnr' tutorials, please visit
    <http://rstudio.github.io/learnr/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rstudio-education.github.io/gradethis/}
  \item \url{https://rstudio.github.io/learnr/}
  \item \url{https://github.com/rstudio-education/gradethis}
  \item Report bugs at \url{https://github.com/rstudio-education/gradethis/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@rstudio.com}

Authors:
\itemize{
  \item Daniel Chen \email{daniel.chen@rstudio.com}
  \item Garrett Grolemund \email{garrett@rstudio.com} [conceptor]
  \item Barret Schloerke \email{barret@rstudio.com}
}

Other contributors:
\itemize{
  \item  RStudio, PBC [copyright holder, funder]
}

}
\keyword{internal}
