% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradethis_demo.R
\name{gradethis_demo}
\alias{gradethis_demo}
\title{Grading Demo}
\usage{
gradethis_demo()
}
\description{
If you are using the RStudio IDE, \code{gradethis_demo()} opens an example learnr
file that demonstrates how to use the grader package to check student code.
}
\details{
The tutorial sets the learnr \code{exercise.checker} option to
\code{grade_learnr()} in the document's setup chunk.

It then uses three different exercise checking methods:
\code{\link[=grade_result]{grade_result()}}, \code{\link[=grade_conditions]{grade_conditions()}}, and \code{\link[=grade_code]{grade_code()}}.

To use a checking method, follow the exercise chunk with a chunk whose label
matches the label of the exercise chunk (ex: \code{myexercise}) but includes the suffix
\code{-check} (ex: \code{myexercise-check}). Call any checking method in that chunk.

To ensure that checking method can provide informative feedback,
you may provide custom \code{correct} and \code{incorrect} messages.

If you are not using RStudio IDE, you can access the demo file at
\code{system.file("extdata", "grading-demo/grading-demo.Rmd", package = "grader")}.
}
