% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradethis_equal.R
\name{gradethis_equal}
\alias{gradethis_equal}
\alias{gradethis_equal.default}
\title{Compare the values of two objects to check whether they are equal}
\usage{
gradethis_equal(x, y, ...)

\method{gradethis_equal}{default}(x, y, tolerance, ...)
}
\arguments{
\item{x, y}{Two objects to compare}

\item{...}{Additional arguments passed to methods}

\item{tolerance}{If non-\code{NULL}, used as threshold for ignoring small
floating point difference when comparing numeric vectors. Setting to
any non-\code{NULL} value will cause integer and double vectors to be compared
based on their values, rather than their types.

It uses the same algorithm as \code{\link[=all.equal]{all.equal()}}, i.e., first we generate
\code{x_diff} and \code{y_diff} by subsetting \code{x} and \code{y} to look only locations
with differences. Then we check that
\code{mean(abs(x_diff - y_diff)) / mean(abs(y_diff))} (or just
\code{mean(abs(x_diff - y_diff))} if \code{y_diff} is small) is less than
\code{tolerance}.}
}
\value{
A \link{logical} value of length one, or an internal gradethis error.
}
\description{
Compare the values of two objects to check whether they are equal
}
\section{Methods (by class)}{
\itemize{
\item \code{gradethis_equal(default)}: The default comparison method, which uses \link[waldo:compare]{waldo::compare}

}}
\examples{
gradethis_equal(mtcars[mtcars$cyl == 6, ], mtcars[mtcars$cyl == 6, ])
gradethis_equal(mtcars[mtcars$cyl == 6, ], mtcars[mtcars$cyl == 4, ])
}
