% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradethis_error_checker.R
\name{gradethis_error_checker}
\alias{gradethis_error_checker}
\title{An error checking function for use with learnr}
\usage{
gradethis_error_checker(
  ...,
  hint = getOption("gradethis.fail.hint", TRUE),
  message = getOption("gradethis.error_checker.message", NULL),
  encourage = getOption("gradethis.fail.encourage", FALSE)
)
}
\arguments{
\item{...}{Ignored but included for future compatibility.}

\item{hint}{Include a code feedback hint with the failing message? This
argument only applies to \code{fail()} and \code{fail_if_equal()} and the message is
added using the default options of \code{\link[=give_code_feedback]{give_code_feedback()}} and
\code{\link[=maybe_code_feedback]{maybe_code_feedback()}}. The default value of \code{hint} can be set using
\code{\link[=gradethis_setup]{gradethis_setup()}} or the \code{gradethis.fail.hint} option.}

\item{message}{The feedback message when an error occurred and no solution is
provided for the exercise. May reference \code{.error} or any of the
\link{grade_this-objects}. The default value is set by \code{\link[=gradethis_setup]{gradethis_setup()}}.}

\item{encourage}{Include a random encouraging phrase with
\code{\link[=random_encouragement]{random_encouragement()}}? The default value of \code{encourage} can be set
using \code{\link[=gradethis_setup]{gradethis_setup()}} or the \code{gradethis.fail.encourage} option.}
}
\value{
A checking function compatible with \code{\link[=gradethis_exercise_checker]{gradethis_exercise_checker()}}.
}
\description{
\pkg{learnr} uses the checking code in \code{exercise.error.check.code} when the
user's submission produces an error during evaluation.
\code{gradethis_error_checker()} provides default error checking suitable for most
situations where an error was \emph{not expected}.

If a solution for the exercise is available, the user's submission will be
compared to the example solution and the message to the student will include
code feedback. Otherwise, the error message from R is returned.

If you \emph{are expecting} the user to submit code that throws an error, use the
\verb{*-error-check} chunk to write custom grading code that validates that the
correct error was created.
}
\examples{
# The default error checker is run on an exercise that produces an error.
# In the following example, the object `b` is not defined.

# This is the error that the user's submission creates:
tryCatch(
  b,
  error = function(e) message(e$message)
)

# If you haven't provided a model solution:
gradethis_error_checker()(mock_this_exercise(b))

# If a model solution is available:
gradethis_error_checker()(mock_this_exercise(b, a))
}
\seealso{
\code{\link[=gradethis_setup]{gradethis_setup()}}, \code{\link[=gradethis_exercise_checker]{gradethis_exercise_checker()}}
}
