% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradethis_exercise_checker.R
\name{gradethis_exercise_checker}
\alias{gradethis_exercise_checker}
\title{A checker function to use with \pkg{learnr}}
\usage{
gradethis_exercise_checker(
  label = NULL,
  solution_code = NULL,
  user_code = NULL,
  check_code = NULL,
  envir_result = NULL,
  evaluate_result = NULL,
  envir_prep = NULL,
  last_value = NULL,
  stage = NULL,
  ...,
  solution_eval_fn = NULL
)
}
\arguments{
\item{label}{Label for exercise chunk}

\item{solution_code}{Code provided within the "-solution" chunk for the
exercise.}

\item{user_code}{R code submitted by the user}

\item{check_code}{Code provided within the "-check" (or "-code-check") chunk
for the exercise.}

\item{envir_result}{The R environment after the execution of the chunk.}

\item{evaluate_result}{The return value from the \code{evaluate::evaluate}
function.}

\item{envir_prep}{A copy of the R environment before the execution of the
chunk.}

\item{last_value}{The last value from evaluating the user's exercise
submission.}

\item{stage}{The current stage of exercise checking.}

\item{...}{Extra arguments supplied by learnr}

\item{solution_eval_fn}{A function taking solution \code{code} and an \code{envir}
(an environment equivalent to \code{envir_prep}) and that should return the
value of the evaluated \code{code}. This callback function allows grading
authors to write custom solution evaluation functions for non-R exercise
engines. The result of the evaluated code should be an R object that will
be accessible to the grading code in \link{.solution} or \link{.solution_all}.}
}
\value{
Returns a feedback object suitable for \pkg{learnr} tutorials with
the results of the exercise grading code.
}
\description{
For exercise checking, \pkg{learnr} tutorials require a function that
\pkg{learnr} can use in the background to run the code in each "-check"
chunk and to format the results into a format that \pkg{learnr} can display.
To enable exercise checking in your learnr tutorial, attach \pkg{gradethis}
with \code{library(gradethis)}, or call \code{gradethis_setup()} in the setup chunk
of your tutorial. See \code{gradethis_demo()} to see an example learnr document
that uses \code{gradethis_exercise_checker()}.
}
\examples{
\dontrun{
gradethis_demo()
}

}
\seealso{
\code{\link[=gradethis_setup]{gradethis_setup()}}, \code{\link[=grade_this]{grade_this()}}, \code{\link[=grade_this_code]{grade_this_code()}}
}
