% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradethis_setup.R
\name{gradethis_setup}
\alias{gradethis_setup}
\title{Setup gradethis for use within learnr}
\usage{
gradethis_setup(
  pass = NULL,
  fail = NULL,
  ...,
  code_correct = NULL,
  code_incorrect = NULL,
  maybe_code_feedback = NULL,
  pipe_warning = NULL,
  allow_partial_matching = NULL,
  exercise.checker = gradethis_exercise_checker,
  exercise.timelimit = NULL,
  exercise.error.check.code = NULL,
  fail_code_feedback = NULL
)
}
\arguments{
\item{pass}{Default message for \code{\link[=pass]{pass()}}. Sets \code{options("gradethis.pass")}}

\item{fail}{Default message for \code{\link[=fail]{fail()}}. Sets \code{options("gradethis.fail")}}

\item{...}{
  Arguments passed on to \code{\link[learnr:tutorial_options]{learnr::tutorial_options}}
  \describe{
    \item{\code{exercise.cap}}{Caption for exercise chunk (defaults to the engine's icon or the combination of the engine and \code{" code"}).}
    \item{\code{exercise.eval}}{Whether to pre-evaluate the exercise so the reader can
see some default output (defaults to \code{FALSE}).}
    \item{\code{exercise.lines}}{Lines of code for exercise editor (defaults to the
number of lines in the code chunk).}
    \item{\code{exercise.completion}}{Use code completion in exercise editors.}
    \item{\code{exercise.diagnostics}}{Show diagnostics in exercise editors.}
    \item{\code{exercise.startover}}{Show "Start Over" button on exercise.}
    \item{\code{exercise.reveal_solution}}{Whether to reveal the exercise solution if
a solution chunk is provided.}
  }}

\item{code_correct}{Default \code{correct} message for \code{\link[=grade_this_code]{grade_this_code()}}. If
unset, \code{\link[=grade_this_code]{grade_this_code()}} falls back to the value of the \code{gradethis.pass}
option. Sets \code{options("gradethis.code_correct")}.}

\item{code_incorrect}{Default \code{incorrect} message for \code{\link[=grade_this_code]{grade_this_code()}}. If
unset \code{\link[=grade_this_code]{grade_this_code()}} falls back to the value of the \code{gradethis.fail}
option. Sets \code{options("gradethis.code_incorrect")}.}

\item{maybe_code_feedback}{Logical \code{TRUE} or \code{FALSE} to determine whether
\code{\link[=maybe_code_feedback]{maybe_code_feedback()}} should return code feedback, where if \code{FALSE},
\code{\link[=maybe_code_feedback]{maybe_code_feedback()}} will return an empty string.
\code{\link[=maybe_code_feedback]{maybe_code_feedback()}} is used in the default messages when \code{\link[=pass]{pass()}} or
\code{\link[=fail]{fail()}} are called without any arguments, which are set by the \code{pass} or
\code{fail} arguments of \code{\link[=gradethis_setup]{gradethis_setup()}}.}

\item{pipe_warning}{The default message used in \code{\link[=pipe_warning]{pipe_warning()}}. Sets
\code{options("gradethis.pipe_warning")}.}

\item{allow_partial_matching}{Logical \code{TRUE} or \code{FALSE} to determine whether
partial matching is allowed in \code{grade_this_code()}. Sets
\code{options("gradethis.allow_partial_matching")}.}

\item{exercise.checker}{Function used to check exercise answers
(e.g., \code{gradethis::grade_learnr()}).}

\item{exercise.timelimit}{Number of seconds to limit execution time to
(defaults to \code{30}).}

\item{exercise.error.check.code}{A string containing R code to use for checking
code when an exercise evaluation error occurs (e.g., \code{"gradethis::grade_code()"}).}

\item{fail_code_feedback}{Deprecated. Use \code{maybe_code_feedback}.}
}
\value{
Invisibly returns the global options as they were prior to setting
them with \code{gradethis_setup()}.
}
\description{
Use this function to change the default options suggested by gradethis. This
function also describes in detail each of the global options available for
customization in the gradethis package. Note that you most likely do not want
to change the defaults values for the learnr tutorial options that are
prefixed with \code{exercise.}. Each of the gradethis-specific arguments sets a
global option with the same name, prefixed with \code{gradethis.}. For example,
\code{pass} sets \code{gradethis.pass}.
}
\examples{
# Not run in package documentation because this function changes global opts
if (FALSE) {
  old_opts <- gradethis_setup(
    pass = "Great work!", 
    fail = "{random_encouragement()}"
  )
}

# Use getOption() to see the default value
getOption("gradethis.pass")
getOption("gradethis.maybe_code_feedback")

}
\seealso{
\code{\link[=gradethis_exercise_checker]{gradethis_exercise_checker()}}
}
