% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_env.R
\name{merge_environnements}
\alias{merge_environnements}
\title{merge two environnements in one}
\usage{
merge_environnements(...)
}
\arguments{
\item{...}{environnements}
}
\value{
an environnement
}
\description{
merge two environnements in one
}
\details{
in case of name conflit le last environnement win
}
\examples{
e1 <- new.env()
e1$a <- "a from e1"
e2 <- new.env()
e2$a <- "a from e2"
e2$b <- "b from e2"

new <- merge_environnement(e1 = e1, e2 = e2)

merge_environnement(e1, e2)$a # "a from e2"
merge_environnement(e2, e1)$a # "a from e1"
}
