% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass_fail_if.R
\name{pass_if}
\alias{pass_if}
\alias{fail_if}
\title{Condition definition}
\usage{
pass_if(x, message = NULL)

fail_if(x, message = NULL)
}
\arguments{
\item{x}{A formula, function, or value, that returns \code{TRUE} or \code{FALSE}.
When comparing objects that are greater than length 1 (e.g., vectors, dataframes, matricies, etc)
A boolean vector will be returned if the user uses \code{==}, not a single boolean value.
\code{grader} will run the vector through \code{all(..., na.rm = TRUE)} to check for the boolean value.
It is advised that the user use \code{identical()} instead of \code{==} in this case.}

\item{message}{chracter string for message returned}
}
\description{
Wrapper around \code{\link{condition}} that sets the proper \code{correct} argument.
This allows the instructor to specify just the comparison code, \code{x},
and \code{message}.
}
\section{Functions}{
\itemize{
\item \code{pass_if}: a condition that if matched means the student provided result is correct

\item \code{fail_if}: a condition that if matched means the student provided result is incorrect
}}

\examples{
\dontrun{grading_demo()}

check_result(
  pass_if(~ identical(.result, 5), "This is a correct message"),
          grader_args = list(),
          learnr_args = list(last_value = 5, envir_prep = new.env())
)

check_result(
  pass_if(function(x) identical(x, 5), "This is a correct message"),
  learnr_args = list(last_value = 5)
)

check_result(
  pass_if(5, "This is a correct message"),
          learnr_args = list(last_value = 5)
)

check_result(
  fail_if(5, "You were supposed to do this other thing!"),
          pass_if(~ TRUE, "should never reach here"),
          learnr_args = list(last_value = 5)
)
}
