% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass_fail_condi.R
\name{pass_if}
\alias{pass_if}
\title{Pass if condition matches}
\usage{
pass_if(x, message)
}
\arguments{
\item{x}{a formula, function, or value, that returns \code{TRUE} or \code{FALSE}}

\item{message}{chracter string for message returned}
}
\description{
Pass if condition matches
}
