% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_solutions.R
\name{print.solutions_test_full}
\alias{print.solutions_test_full}
\title{Print method for solutions test}
\usage{
\method{print}{solutions_test_full}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\description{
Print method for solutions test
}
