% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{random_encouragement}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{learnr}{\code{\link[learnr:random_praise]{random_encouragement}}}
}}

