% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{result}
\alias{result}
\title{Result}
\usage{
result(x, message = NULL, correct = FALSE)
}
\arguments{
\item{x}{object graded or being compared against}

\item{message}{possible message value to be displayed}

\item{correct}{a boolean that determines if the result is a correct result}
}
\description{
The \code{result} method returns an object containing information about what has been graded or what could be graded against.
}
\examples{
result(1, "Custom message for value 1.")
result(2, "Custom message for value 2.", correct = TRUE)
result(3, "Custom message for value 3.")
result(4, "Custom message for value 4.", correct = TRUE)
}
