% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_result.R
\name{results}
\alias{results}
\title{Possible results}
\usage{
results(...)
}
\arguments{
\item{...}{a group of \code{\link{result}} output objects.}
}
\value{
A \code{list} with class \code{"grader_result"} containing \code{\link{result}} objects.
}
\description{
A list of possible results to check against.  Each argument supplied must contain output from \code{\link{result}}.
}
\examples{
results(
  result(1, "Custom message for value 1."),
  result(2, "Custom message for value 2.", correct = TRUE),
  result(3, "Custom message for value 3."),
  result(4, "Custom message for value 4.", correct = TRUE)
)
}
\seealso{
\code{\link{result}}
}
