% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strict_check.R
\name{strict_check}
\alias{strict_check}
\title{strict_check}
\usage{
strict_check(success = "Correct!", solution = NULL, user = NULL)
}
\arguments{
\item{success}{A character string to display if the student answer matches
the solution code}

\item{solution}{(Optional) solution code surrounded by \code{quote()},
\code{rlang::quo()}, or provided as a character string.}

\item{user}{(Optional) student code to check against the solution surrounded
by \code{quote()}, \code{rlang::quo()}, or provided as a character string.}
}
\value{
(character) A message. If the student answer differs from the
  solution code, the message will describe the first way that the answer
  differs, and it will ask the student to try again. If the answer matches
  the solution code, the message will be the content of the \code{success}
  argument.
}
\description{
Strict exercise checking
}
\details{
\code{strict_check()} compares user code to a solution (i.e. model code) and
describes the first way that the user code differs. If the user code exactly
matches the solution, \code{strict_check()} returns a customizable success
message.

\code{strict_check()} provides a *strict* check in that the user code must
exactly match the solution. It is not enough for the user code to be
equivalent to the solution code (e.g. to return the same result as the
solution).

You can provide solution code for \code{strict_check()} to use in two ways:

1. Pass code as a character string or a quoted expression to the solution
argument of \code{strict_check()}

2. Make a "-solution" code chunk for the exercise to be checked in a learnr
document. There is no need to supply a solution argument for
\code{strict_check()} if you call it from the "-check" chunk of the same
exercise. Likewise, there is no need to supply a user argument when you call
\code{strict_check()} from a learnr document (learnr will provide the code
that the student submits when it runs \code{strict_check()}.

For best results, name all arguments provided in the solution code.
}
\examples{
\dontrun{grading_demo()}
}
