% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_result.R
\name{test}
\alias{test}
\title{TODO document with 'tests' documentation}
\usage{
test(message, fn)
}
\arguments{
\item{message}{Message to report back if the test throws an error.}

\item{fn}{function to execute against the user solution.
If the test fails, it should throw an error to display the \code{message} provided.}
}
\description{
TODO document with 'tests' documentation
}
