% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_result.R
\name{test_result}
\alias{test_result}
\title{Test the result of exercise code}
\usage{
test_result(tests,
  correct = "{num_correct}/{num_total} correct! {random_praise()}",
  incorrect = paste0("{num_correct}/{num_total} correct. ",
  "Fix this first: '{errors[1]}'. ", "{random_encourage()}"),
  empty_msg = "I did not notice a result. Does your code return one?",
  ..., user = NULL)
}
\arguments{
\item{tests}{A \code{\link{tests}} object that contains all \code{\link{test}} functions to check the user's result.}

\item{correct}{A character string to display if all tests pass.
  This character string will be run through \code{glue::\link[glue]{glue_data}} with
\itemize{
  \item \code{num_correct}: Number of correct tests. (Equals \code{num_total})
  \item \code{num_total}: Number of tests
  \item \code{errors}: Vector of errors found. (\code{NULL})
}}

\item{incorrect}{A character string to display if at least one test fails.
  This character string will be run through \code{glue::\link[glue]{glue_data}} with
\itemize{
  \item \code{num_correct}: Number of correct tests
  \item \code{num_total}: Number of tests
  \item \code{errors}: Vector of errors found
}}

\item{empty_msg}{A character string to display as a message if the user code is NULL.}

\item{...}{ignored}

\item{user}{(Optional) student code to check against the \code{results} surrounded
by \code{quote()}, \code{rlang::quo()}, or provided as a character string.}
}
\value{
a \code{grader_result} structure from \code{\link{result}} containing a formatted \code{correct} or \code{incorrect} message.
}
\description{
\code{test_result()} executes tests against the final result of the user code.
If a test throws an error, the test fails and the submitted answer will be marked incorrect.
}
\examples{
\dontrun{grading_demo()}
}
\seealso{
\code{tests}, \code{test}
}
