% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_solutions.R
\name{test_solutions}
\alias{test_solutions}
\title{Test Solutions}
\usage{
test_solutions(file, show.answers = FALSE, .params = NULL)
}
\arguments{
\item{file}{The filepath to a learnr tutorial}
}
\description{
Test Solutions
}
