% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpipe.R
\name{unpipe}
\alias{unpipe}
\title{unpipe}
\usage{
unpipe(code)
}
\arguments{
\item{code}{a quoted piece of code}
}
\description{
Unpipe one layer of piped code.
}
\details{
\code{unpipe()} removes one call to \%>\% from an expression,
reformatting the result as a nested function call. If the code does not
contain a top-level call to \%>\%, \code{unpipe()} returns it as is.
}
