% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_object.R
\name{user_object}
\alias{user_object}
\alias{user_object_get}
\alias{user_object_exists}
\alias{user_object_list}
\title{Functions for interacting with objects created by student code}
\usage{
user_object_get(
  x,
  mode = "any",
  envir = .envir_result,
  ...,
  gradethis_env = parent.frame()
)

user_object_exists(
  x,
  mode = "any",
  envir = .envir_result,
  ...,
  gradethis_env = parent.frame()
)

user_object_list(
  mode = "any",
  envir = .envir_result,
  exclude_envir = .envir_prep,
  ...,
  gradethis_env = parent.frame()
)
}
\arguments{
\item{x}{An object name, given as a quoted \link{character} string.}

\item{mode}{character specifying the \code{\link{mode}} of objects to
    consider.  Passed to \code{\link{exists}} and \code{\link{get}}.}

\item{envir}{An \link{environment} in which to search for objects.
Defaults to \code{\link{.envir_result}}.}

\item{exclude_envir}{An \link{environment}.
Objects that appear in both \code{envir} and \code{exclude_envir} will be excluded.
Defaults to \code{\link{.envir_prep}}.
Use \code{exclude_envir = NULL} to include all objects.}

\item{...}{Additional arguments passed to underlying functions:
\itemize{
\item For \code{user_object_exists()}, \code{\link[=exists]{exists()}}
\item For \code{user_object_get()}, \code{\link[=get]{get()}}
\item For \code{user_object_list()}, \code{\link[utils:ls_str]{ls.str()}}
}}

\item{gradethis_env}{The \link{environment} from which to retrieve
\code{\link{.envir_result}} and \code{\link{.envir_prep}}.
Most users of \pkg{gradethis} will not need to use this argument.}
}
\value{
For \code{user_object_get()}, the object.
If the object is not found, an error.

For \code{user_object_exists()}, a \code{\link{TRUE}}/\code{\link{FALSE}} value.

For \code{user_object_list()}, a \link{character} vector
giving the names of the objects in \code{envir}.
}
\description{
Functions for interacting with objects created by student code
}
\examples{
user_code <- rlang::expr({
  # ```{r example}
  x <- 1
  y <- list(1, 2, 3)
  z <- function() print("Hello World!")
  # ```
})

exercise <- mock_this_exercise(!!user_code)
.envir_result <- exercise$.envir_result
.envir_prep <- exercise$.envir_prep

user_object_list()
user_object_exists("x")
user_object_get("y")

# Use `mode` to find only objects of a certain type ----
user_object_list(mode = "numeric")
user_object_list(mode = "list")
user_object_list(mode = "function")

user_object_exists("x", mode = "numeric")
user_object_exists("y", mode = "numeric")

user_object_get("z", mode = "function")

# By default, `user_object_list()` ignores objects created by setup chunks ----
setup_code <- rlang::expr({
  # ```{r example-setup}
  exercise_data <- mtcars
  # ```
})

exercise <- mock_this_exercise(!!user_code, setup_exercise = !!setup_code)
.envir_result <- exercise$.envir_result
.envir_prep <- exercise$.envir_prep

user_object_list()

## You can disable this by setting `exclude_envir = NULL` ----
user_object_list(exclude_envir = NULL)
}
