% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_tutorial.R
\name{view_tutorial}
\alias{view_tutorial}
\title{View tutorial}
\usage{
view_tutorial(name, package)
}
\arguments{
\item{name}{A character string. The name of a tutorial saved in a package.}

\item{package}{A character string. The name of the package that contains a tutorial.}
}
\description{
\code{view_tutorial()} opens a tutorial in the RStudio viewer window. You can
place it in a .Rprofile file to cause a project to open with a tutorial
visible. This makes a convenient way to guide students through sophisticated
case studies, or to transition students from the learnr environment to the
IDE. However, it only works for teaching purposes if students can still use
the IDE command line. How to open the tutorial in a second R process?
}
