% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_exercise.R
\name{with_exercise}
\alias{with_exercise}
\title{Run an expression as if it were in an exercise's \code{grade_this()} block}
\usage{
with_exercise(exercise, expr)
}
\arguments{
\item{exercise}{An exercise, as created by \code{\link[=mock_this_exercise]{mock_this_exercise()}}}

\item{expr}{An unquoted expression}
}
\value{
The value of \verb{grade_this(<expr>)(exercise)}
}
\description{
This function is not intended to be used within grading code,
but may be helpful for testing grading code.
}
\examples{
exercise <- mock_this_exercise(.user_code = "2", .solution_code = "1 + 1")

with_exercise(exercise, pass_if_equal())
with_exercise(exercise, fail_if_code_feedback())
}
