% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_result.R
\name{check_result}
\alias{check_result}
\title{Check result of exercise code}
\usage{
check_result(...,
  correct = "{paste0(random_praise(), if (nchar(message) > 0) \\" \\", message)}",
  incorrect = "{paste0(message, if(nchar(message) > 0) \\" \\", random_encourage())}",
  grader_args = list(), learnr_args = list())
}
\arguments{
\item{...}{ignored}

\item{correct}{A character string to display if the student answer matches
a known answer.
This character string will be run through \code{glue::\link[glue]{glue_data}} with \code{list(correct = TRUE, message = "<result message>")}. where message is the matched result message.}

\item{incorrect}{A character string to display if the student answer matches
a known answer.
This character string will be run through \code{glue::\link[glue]{glue_data}} with \code{list(correct = FALSE, message = "<result message>")}. where message is the matched result message.}

\item{grader_args}{A list of parameters passed to \code{grader} functions (provided by \code{grade_learnr}).
This can be student code to check against the \code{results} surrounded
by \code{quote()}, \code{rlang::quo()}, or provided as a character string.
For example, \code{user_quo = quote(1)}}

\item{learnr_args}{A list of parameters passed to \code{learnr} (provided by \code{grade_learnr}), for example, \code{last_value = 1}}
}
\value{
a \code{grader_graded} structure from \code{\link{result}} containing a formatted \code{correct} or \code{incorrect} message and whether or not a match was found, \code{matched}.
}
\description{
\code{check_result()} compares the final result of the user code to known \code{results}.
If the user result exactly matches a known \code{result}, \code{check_result}
returns the matching message value.
}
\examples{
\dontrun{grading_demo()}
}
