% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_result.R
\name{check_result}
\alias{check_result}
\title{Check result of exercise code}
\usage{
check_result(results, correct = paste0("{random_praise()}",
  "{paste0(if (!is.null(message)) paste0(\\" \\", message))}"),
  incorrect = paste0("{paste0(if(!is.null(message)) paste0(message, \\" \\"))}",
  "{random_encourage()}"),
  empty_msg = "I did not notice a result. Does your code return one?",
  ..., user = NULL)
}
\arguments{
\item{results}{A \code{\link{results}} object that contains possible \code{\link{result}} values to compare against.}

\item{correct}{A character string to display if the student answer matches
a known answer.
This character string will be run through \code{glue::\link[glue]{glue_data}} with \code{list(correct = TRUE, message = "<result message>")}. where message is the matched result message.}

\item{incorrect}{A character string to display if the student answer matches
a known answer.
This character string will be run through \code{glue::\link[glue]{glue_data}} with \code{list(correct = FALSE, message = "<result message>")}. where message is the matched result message.}

\item{empty_msg}{A character string to display as a message if the user code is NULL.}

\item{...}{ignored}

\item{user}{(Optional) student code to check against the \code{results} surrounded
by \code{quote()}, \code{rlang::quo()}, or provided as a character string.}
}
\value{
a \code{grader_result} structure from \code{\link{result}} containing a formatted \code{correct} or \code{incorrect} message.
}
\description{
\code{check_result()} compares the final result of the user code to known \code{results}.
If the user result exactly matches a known \code{result}, \code{check_result}
returns the matching message value.
}
\examples{
\dontrun{grading_demo()}
}
