% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_result.R
\name{check_result}
\alias{check_result}
\title{Check result of exercise code}
\usage{
check_result(..., correct = NULL, incorrect = NULL,
  grader_args = list(), learnr_args = list(),
  glue_correct = getOption("gradethis_glue_correct"),
  glue_incorrect = getOption("gradethis_glue_incorrect"))
}
\arguments{
\item{...}{ignored}

\item{correct}{A character string to display if the student answer matches a known answer.}

\item{incorrect}{A character string to display if the student answer matches
a known answer.
This character string will be run through
\code{glue::\link[glue]{glue_data}} with
\code{list(correct = FALSE, message = "<result message>")}.
where message is the matched result message.}

\item{grader_args}{A list of parameters passed to \code{grader} functions (provided by \code{grade_learnr}).
This contains: \describe{
 \item{\code{user_quo}}{Quoted R code submitted by the user.  Ex: \code{rlang::\link[rlang]{quo}(1)} }
 \item{\code{solution_quo}}{[Optional] Quoted solution R code provided by the \code{*-solution} chunk for an exercise.}
}}

\item{learnr_args}{A list of all parameters passed to \code{\link{grade_learnr}} by \code{learnr}.
See \url{https://rstudio.github.io/learnr/exercises.html#exercise_checking} for more details.}

\item{glue_correct}{A glue string that returns the final correct message displayed.
Defaults to getOption("gradethis_glue_correct").}

\item{glue_incorrect}{A glue string that returns the final incorrect message displayed.
Defaults to getOption("gradethis_glue_incorrect").}
}
\value{
a \code{grader_graded} structure from either
  \code{\link{pass_if}} or \code{\link{fail_if}} containing a formatted
  \code{correct} or \code{incorrect} message and whether or not a match was found.
}
\description{
\code{check_result()} compares the final result of the user code to known \code{results}.
If the user result exactly matches a known \code{result}, \code{check_result}
returns the matching message value.
}
\examples{
\dontrun{grading_demo()}
}
