% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R, R/pass_fail_if.R
\name{condition}
\alias{condition}
\alias{pass_if}
\alias{fail_if}
\title{Condition object}
\usage{
condition(x, message, correct)

pass_if(x, message = NULL)

fail_if(x, message = NULL)
}
\arguments{
\item{x}{A formula, function, or value, that returns \code{TRUE} or \code{FALSE}.
When comparing objects that are greater than length 1
(e.g., vectors, dataframes, matricies, etc)
A boolean vector will be returned if the user uses \code{==}, not a single boolean value.
\code{gradethis} will run the vector through
\code{all(..., na.rm = TRUE)} to check for the boolean value.
It is advised that the user use \code{identical()} instead of \code{==} in this case.}

\item{message}{character string for message returned (usually passed in from
\code{\link[=pass_if]{pass_if()}} or \code{\link[=fail_if]{fail_if()}}}

\item{correct}{logical whether the condition is the correct answer}
}
\value{
a \code{grader_condition} object that contains
the expression \code{x},
the message \code{message},
whether or not the expression is the correct answer or not, \code{correct},
the type of expression (formula, function, or value), \code{type}
}
\description{
Captures what the student passes into \code{\link[=pass_if]{pass_if()}} or \code{\link[=fail_if]{fail_if()}},
figures out what type of object was passed into \code{x},
and returns a \code{grader_condition} object that will be passed into
\code{\link[=evaluate_condition]{evaluate_condition()}}.

\code{pass_if} and \code{fail_if} are wrappers around \code{\link[=condition]{condition()}}
that sets the proper \code{correct} argument.
This allows the instructor to specify just the comparison code, \code{x},
and \code{message}.
}
\section{Functions}{
\itemize{
\item \code{condition}: A condition object that contains the expression, message,
correct status, and condition type

\item \code{pass_if}: a condition that if matched means the student provided result is correct

\item \code{fail_if}: a condition that if matched means the student provided result is incorrect
}}

\examples{
\dontrun{gradethis_demo()}

condition(~ identical(x = .result, 5), message = "Correct", correct = TRUE)
grade_result(
  pass_if(~ identical(.result, 5), "This is a correct message"),
          grader_args = list(),
          learnr_args = list(last_value = 5, envir_prep = new.env())
)

grade_result(
  pass_if(function(x) identical(x, 5), "This is a correct message"),
  learnr_args = list(last_value = 5)
)

grade_result(
  pass_if(5, "This is a correct message"),
          learnr_args = list(last_value = 5)
)

grade_result(
  fail_if(5, "You were supposed to do this other thing!"),
          pass_if(~ TRUE, "should never reach here"),
          learnr_args = list(last_value = 5)
)
}
\seealso{
\code{\link[=pass_if]{pass_if()}}, \code{\link[=fail_if]{fail_if()}}, and \code{\link[=condition]{condition()}}
}
