% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass_fail_condi.R
\name{evaluate_condition}
\alias{evaluate_condition}
\title{Evaluates a condition}
\usage{
evaluate_condition(condi, grader_args, learnr_args)
}
\arguments{
\item{condi}{a \code{grader} \code{\link{condition}} object}

\item{grader_args}{at minimum, a list that just contains the value for \code{solution_quo}}

\item{learnr_args}{at minimum, a list that just contains the value for \code{envir_prep}}
}
\value{
a \code{graded} value if \code{condi$x} is \code{TRUE} or
  \code{NULL} if \code{condi$x} is \code{FALSE}
}
\description{
Evaluates a condition
}
