% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_condition.R
\name{evaluate_condition}
\alias{evaluate_condition}
\title{Evaluates a condition}
\usage{
evaluate_condition(condition, grader_args, learnr_args)
}
\arguments{
\item{condition}{a \code{\link[=condition]{condition()}} object}

\item{grader_args}{at minimum, a list that just contains the value for \code{solution_quo}}

\item{learnr_args}{at minimum, a list that just contains the value for \code{envir_prep}}
}
\value{
a \code{\link[=graded]{graded()}} value if \code{condi$x} is \code{TRUE} or
\code{NULL} if \code{condi$x} is \code{FALSE}
}
\description{
Evaluates the \code{\link[=condition]{condition()}} object to return a \code{\link[=graded]{graded()}} value.
}
\examples{
 condi_formula_t <- condition(~ identical(.result, 5),
                              message = "my correct message",
                              correct = TRUE)
 grader_args <- list()
 learnr_args <- list(last_value = quote(5), envir_prep = new.env())
 evaluate_condition(condi_formula_t, grader_args, learnr_args)
}
