% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass_fail_condi.R
\name{fail_if}
\alias{fail_if}
\title{Fail if condition matches}
\usage{
fail_if(x, message)
}
\arguments{
\item{x}{a formula, function, or value, that returns \code{TRUE} or \code{FALSE}}

\item{message}{chracter string for message returned}
}
\description{
Fail if condition matches
}
