% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graded.R
\name{fail_if_code_feedback}
\alias{fail_if_code_feedback}
\title{Signal a failing grade if mistakes are detected in the submitted code}
\usage{
fail_if_code_feedback(
  user_code = missing_arg(),
  solution_code = missing_arg(),
  message = NULL,
  ...,
  env = parent.frame(),
  hint = TRUE,
  encourage = getOption("gradethis.fail.encourage", FALSE),
  allow_partial_matching = getOption("gradethis.allow_partial_matching", TRUE)
)
}
\arguments{
\item{user_code}{String containing user or solution code. For
ease of use in \code{grade_this()}, \code{.user_code} or \code{.solution_code} are by
default retrieved from the calling environment.}

\item{solution_code}{String containing user or solution code. For
ease of use in \code{grade_this()}, \code{.user_code} or \code{.solution_code} are by
default retrieved from the calling environment.}

\item{message}{A character string of the message to be displayed. In all
grading helper functions other than \code{\link[=graded]{graded()}}, \code{message} is a template
string that will be processed with \code{\link[glue:glue]{glue::glue()}}.}

\item{...}{
  Arguments passed on to \code{\link[=graded]{graded}}
  \describe{
    \item{\code{correct}}{A logical value of whether or not the checked code is correct.}
    \item{\code{type}}{The \code{type} and \code{location} of the feedback object
provided to \pkg{learnr}. See
\url{https://rstudio.github.io/learnr/exercises.html#Custom_checking} for more
details.

\code{type} may be one of "auto", "success", "info", "warning", "error", or
"custom".

\code{location} may be one of "append", "prepend", or "replace".}
    \item{\code{location}}{The \code{type} and \code{location} of the feedback object
provided to \pkg{learnr}. See
\url{https://rstudio.github.io/learnr/exercises.html#Custom_checking} for more
details.

\code{type} may be one of "auto", "success", "info", "warning", "error", or
"custom".

\code{location} may be one of "append", "prepend", or "replace".}
  }}

\item{env}{Environment used to standardise formals of the user and solution
code. Defaults to retrieving \code{.envir_prep} from the calling environment. If
not found, the \code{\link[=parent.frame]{parent.frame()}} will be used.}

\item{hint}{Include a code feedback hint with the failing message? This
argument only applies to \code{fail()} and \code{fail_if_equal()} and the message is
added using the default options of \code{\link[=give_code_feedback]{give_code_feedback()}} and
\code{\link[=maybe_code_feedback]{maybe_code_feedback()}}. The default value of \code{hint} can be set using
\code{\link[=gradethis_setup]{gradethis_setup()}} or the \code{gradethis.fail.hint} option.}

\item{encourage}{Incude a random encouraging phrase with
\code{\link[=random_encouragement]{random_encouragement()}}? The default value of \code{encourage} can be set
using \code{\link[=gradethis_setup]{gradethis_setup()}} or the \code{gradethis.fail.encourage} option.}

\item{allow_partial_matching}{A logical. If \code{FALSE}, the partial matching of
argument names is not allowed and e.g. \code{runif(1, mi = 0)} will return a
message indicating that the full formal name \code{min} should be used. The
default is set via the \code{gradethis.allow_partial_matching} option, or by
\code{\link[=gradethis_setup]{gradethis_setup()}}.}
}
\value{
Signals an incorrect grade with feedback if there are differences
between the submitted user code and the solution code. If solution code is
not available, no grade is returned.
}
\description{
\code{fail_if_code_feedback()} uses \code{\link[=code_feedback]{code_feedback()}} to detect if there are
differences between the user's submitted code and the solution code (if
available). If the exercise does not have an associated solution, or if there
are no detected differences between the user's and the solution code, no
grade is returned.
}
\examples{
# TODO: examples

}
\seealso{
Grading helper functions
}
