% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{get_user_code}
\alias{get_user_code}
\alias{get_solution_code}
\alias{get_code}
\title{Get Code}
\usage{
get_user_code(user = NULL, env = rlang::caller_env())

get_solution_code(solution = NULL, env = rlang::caller_env())

get_code(expr = NULL, name = "<name not provided>",
  env = rlang::caller_env())
}
\arguments{
\item{user, solution, expr}{An expression or quosure to evaluate.}

\item{env}{The environment in which to evaluate \code{expr}. This
environment is not applicable for quosures because they have
their own environments.}

\item{name}{Name to print if a \code{NULL} expression is provided.}
}
\description{
Helper methods around \code{rlang::\link[rlang]{eval_tidy}}
to extract user code and solution code.
}
\section{Functions}{
\itemize{
\item \code{get_user_code}: get student code

\item \code{get_solution_code}: get solution code provided by instructor

\item \code{get_code}: generic get code function
}}

\seealso{
\code{\link{check_result}}, \code{\link{test_result}}, and \code{\link{check_code}}
}
