% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_conditions.R
\name{grade_conditions}
\alias{grade_conditions}
\title{Grade all specified conditions}
\usage{
grade_conditions(
  ...,
  correct = NULL,
  incorrect = NULL,
  grader_args = list(),
  learnr_args = list(),
  glue_correct = getOption("gradethis_glue_correct_test"),
  glue_incorrect = getOption("gradethis_glue_incorrect_test")
)
}
\description{
This function is deprecated, use \code{\link[=grade_result_strict]{grade_result_strict()}} instead.
}
\keyword{internal}
