% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_this.R
\docType{data}
\name{grade_this-objects}
\alias{grade_this-objects}
\alias{.result}
\alias{.user}
\alias{.last_value}
\alias{.solution}
\alias{.user_code}
\alias{.solution_code}
\alias{.envir_prep}
\alias{.envir_result}
\alias{.evaluate_result}
\alias{.label}
\title{Checking environment objects}
\format{
An object of class \code{.result} (inherits from \code{gradethis_placeholder}) of length 0.

An object of class \code{.user} (inherits from \code{.result}, \code{gradethis_placeholder}) of length 0.

An object of class \code{.last_value} (inherits from \code{.result}, \code{gradethis_placeholder}) of length 0.

An object of class \code{.solution} (inherits from \code{gradethis_placeholder}) of length 0.

An object of class \code{.user_code} (inherits from \code{gradethis_placeholder}) of length 0.

An object of class \code{.solution_code} (inherits from \code{gradethis_placeholder}) of length 0.

An object of class \code{.envir_prep} (inherits from \code{gradethis_placeholder}) of length 0.

An object of class \code{.envir_result} (inherits from \code{gradethis_placeholder}) of length 0.

An object of class \code{.evaluate_result} (inherits from \code{gradethis_placeholder}) of length 0.

An object of class \code{.label} (inherits from \code{gradethis_placeholder}) of length 0.
}
\usage{
.result

.user

.last_value

.solution

.user_code

.solution_code

.envir_prep

.envir_result

.evaluate_result

.label
}
\description{
\code{\link[=grade_this]{grade_this()}} allows instructors to determine a grade and to create custom
feedback messages using custom R code. To facilitate evaluating the
exercise, \code{\link[=grade_this]{grade_this()}} makes available a number of objects that can be
referenced within the \code{{ ... }} expression.

All of the objects provided by \code{learnr} to an exercise checking function
are available for inspection. To avoid name collisions with user or
instructor code, the names of these objects all start with \code{.}.
\itemize{
\item \code{.label}: Label for exercise chunk
\item \code{.last_value} The last value from evaluating the user's exercise
submission
\item \code{.solution_code}: A string containing the code provided within the
\verb{*-solution} chunk for the exercise
\item \code{.user_code}: A string containing the code submitted by the user
\item \code{.check_code}: A string containiner the code provided within the
\verb{*-check} (or \verb{*-code-check}) chunk for the exercise
\item \code{.envir_prep}: A copy of the R environment before the execution of the
chunk
\item \code{.envir_result}: The R environment after the execution of the chunk.
\item \code{.evaluate_result}: The return value from the \code{evaluate::evaluate}
function
}

In addition, \pkg{gradethis} has provided some extra objects:
\itemize{
\item \code{.user}, \code{.result}: A direct copy of \code{.last_value} for friendlier naming
\item \code{.solution}: When accessed, will be the result of evaluating the
\code{.solution_code} in a child environment of \code{.envir_prep}
}

As the instructor, you are free to use any logic to determine a student's
grade as long as a \code{\link[=graded]{graded()}} object is signaled. The check code can also
contain \pkg{testthat} expectation code. Failed \pkg{testthat} expectations
will be turned into \code{\link[=fail]{fail()}}ed grades with the corresponding message.
}
\keyword{datasets}
