% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{graded}
\alias{graded}
\title{\code{grader_graded} object
returned by every \code{check_*} function}
\usage{
graded(message, correct)
}
\arguments{
\item{message}{A character string of the message to be displayed.}

\item{correct}{A boolean value of whether or not the checked code is correct.}
}
\description{
\code{grader_graded} object
returned by every \code{check_*} function
}
