% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{gradethis-deprecated}
\alias{gradethis-deprecated}
\alias{grade_feedback}
\alias{grade_conditions}
\alias{random_encourage}
\alias{grade_learnr}
\title{Deprecated}
\usage{
grade_feedback(...)

grade_conditions(
  ...,
  correct = NULL,
  incorrect = NULL,
  grader_args = deprecated(),
  learnr_args = deprecated(),
  glue_correct = getOption("gradethis.glue_correct_test"),
  glue_incorrect = getOption("gradethis.glue_incorrect_test")
)

random_encourage()

grade_learnr(...)
}
\description{
Deprecated
}
\section{Functions}{
\itemize{
\item \code{grade_feedback}: \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Removed
from package

\item \code{grade_conditions}: \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}} Use
\code{\link[=grade_result_strict]{grade_result_strict()}}

\item \code{random_encourage}: \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}} Use
\code{\link[=random_encouragement]{random_encouragement()}}.

\item \code{grade_learnr}: \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}} Use
\code{\link[=gradethis_exercise_checker]{gradethis_exercise_checker()}}.
}}

\keyword{internal}
