% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{gradethis-deprecated}
\alias{gradethis-deprecated}
\alias{grade_feedback}
\title{Deprecated}
\usage{
grade_feedback(...)
}
\description{
Deprecated
}
\section{Functions}{
\itemize{
\item \code{grade_feedback}: Removed from package
}}

\keyword{internal}
