% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradethis-package.R
\docType{package}
\name{gradethis-package}
\alias{gradethis}
\alias{gradethis-package}
\title{gradethis: Automated Feedback for Student Exercises in 'learnr' Tutorials}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Pairing with the 'learnr' R package, 'gradethis' provides multiple methods to grade 'learnr' exercises. To learn more about 'learnr' tutorials, please visit \url{https://rstudio.github.io/learnr/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkgs.rstudio.com/gradethis/}
  \item \url{https://rstudio.github.io/learnr/}
  \item \url{https://github.com/rstudio/gradethis}
  \item Report bugs at \url{https://github.com/rstudio/gradethis/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@rstudio.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})

Authors:
\itemize{
  \item Daniel Chen \email{daniel.chen@rstudio.com} (\href{https://orcid.org/0000-0003-3857-1741}{ORCID})
  \item Garrett Grolemund \email{garrett@rstudio.com} (\href{https://orcid.org/0000-0002-7765-6011}{ORCID}) [conceptor]
  \item Barret Schloerke \email{barret@rstudio.com} (\href{https://orcid.org/0000-0001-9986-114X}{ORCID})
}

Other contributors:
\itemize{
  \item  RStudio, PBC [copyright holder, funder]
}

}
\keyword{internal}
