% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradethis_setup.R
\name{gradethis_setup}
\alias{gradethis_setup}
\title{Setup gradethis for use within learnr}
\usage{
gradethis_setup(
  pass = NULL,
  fail = NULL,
  ...,
  code_correct = NULL,
  code_incorrect = NULL,
  maybe_code_feedback = NULL,
  maybe_code_feedback.before = NULL,
  maybe_code_feedback.after = NULL,
  pass.praise = NULL,
  fail.hint = NULL,
  fail.encourage = NULL,
  pipe_warning = NULL,
  grading_problem.message = NULL,
  grading_problem.type = NULL,
  error_checker.message = NULL,
  allow_partial_matching = NULL,
  exercise.checker = gradethis_exercise_checker,
  exercise.timelimit = NULL,
  compare_timelimit = NULL,
  exercise.error.check.code = NULL,
  fail_code_feedback = NULL
)
}
\arguments{
\item{pass}{Default message for \code{\link[=pass]{pass()}}. Sets \code{options("gradethis.pass")}}

\item{fail}{Default message for \code{\link[=fail]{fail()}}. Sets \code{options("gradethis.fail")}}

\item{...}{
  Arguments passed on to \code{\link[learnr:tutorial_options]{learnr::tutorial_options}}
  \describe{
    \item{\code{exercise.cap}}{Caption for exercise chunk (defaults to the engine's icon or the combination of the engine and \code{" code"}).}
    \item{\code{exercise.eval}}{Whether to pre-evaluate the exercise so the reader can
see some default output (defaults to \code{FALSE}).}
    \item{\code{exercise.lines}}{Lines of code for exercise editor (defaults to the
number of lines in the code chunk).}
    \item{\code{exercise.blanks}}{A regular expression to be used to identify blanks in
submitted code that the user should fill in. If \code{TRUE} (default), blanks
are three or more underscores in a row. If \code{FALSE}, blank checking is not
performed.}
    \item{\code{exercise.completion}}{Use code completion in exercise editors.}
    \item{\code{exercise.diagnostics}}{Show diagnostics in exercise editors.}
    \item{\code{exercise.startover}}{Show "Start Over" button on exercise.}
    \item{\code{exercise.reveal_solution}}{Whether to reveal the exercise solution if
a solution chunk is provided.}
  }}

\item{code_correct}{Default \code{correct} message for \code{\link[=grade_this_code]{grade_this_code()}}. If
unset, \code{\link[=grade_this_code]{grade_this_code()}} falls back to the value of the \code{gradethis.pass}
option. Sets the \code{gradethis.code_correct} option.}

\item{code_incorrect}{Default \code{incorrect} message for \code{\link[=grade_this_code]{grade_this_code()}}. If
unset \code{\link[=grade_this_code]{grade_this_code()}} falls back to the value of the \code{gradethis.fail}
option. Sets the \code{gradethis.code_incorrect} option.}

\item{maybe_code_feedback}{Logical \code{TRUE} or \code{FALSE} to determine whether
\code{\link[=maybe_code_feedback]{maybe_code_feedback()}} should return code feedback, where if \code{FALSE},
\code{\link[=maybe_code_feedback]{maybe_code_feedback()}} will return an empty string.
\code{\link[=maybe_code_feedback]{maybe_code_feedback()}} is used in the default messages when \code{\link[=pass]{pass()}} or
\code{\link[=fail]{fail()}} are called without any arguments, which are set by the \code{pass} or
\code{fail} arguments of \code{\link[=gradethis_setup]{gradethis_setup()}}.}

\item{maybe_code_feedback.before, maybe_code_feedback.after}{Text that should
be added \code{before} or \code{after} the \code{maybe_code_feedback()} output, if any is
returned. Sets the default values of the \code{before} and \code{after} arguments of
\code{\link[=maybe_code_feedback]{maybe_code_feedback()}}.}

\item{pass.praise}{Logical \code{TRUE} or \code{FALSE} to determine whether a praising
phrase should be automatically prepended to any \code{\link[=pass]{pass()}} or
\code{\link[=pass_if_equal]{pass_if_equal()}} messages. Sets the \code{gradethis.pass.praise} option.}

\item{fail.hint}{Logical \code{TRUE} or \code{FALSE} to determine whether an automated
code feedback hint should be shown with a \code{\link[=fail]{fail()}} or \code{\link[=fail_if_equal]{fail_if_equal()}}
message. Sets the \code{gradethis.fail.hint} option.}

\item{fail.encourage}{Logical \code{TRUE} or \code{FALSE} to determine whether an
encouraging phrase should be automatically appended to any \code{\link[=fail]{fail()}} or
\code{\link[=fail_if_equal]{fail_if_equal()}} messages. Sets the \code{gradethis.fail.encourage} option.}

\item{pipe_warning}{The default message used in \code{\link[=pipe_warning]{pipe_warning()}}. Sets the
\code{gradethis.pipe_warning} option.}

\item{grading_problem.message}{The feedback message used when a grading error occurs.
Sets the \code{gradethis.grading_problem.message} option.}

\item{grading_problem.type}{The feedback type used when a grading error occurs.
Must be one of \code{"success"}, \code{"info"}, \code{"warning"} (default), \code{"error"}, or
\code{"custom"}. Sets the \code{gradethis.grading_problem.type} option.}

\item{error_checker.message}{The default message used by gradethis's default
error checker, \code{\link[=gradethis_error_checker]{gradethis_error_checker()}}. Sets the
\code{gradethis.error_checker.message} option.}

\item{allow_partial_matching}{Logical \code{TRUE} or \code{FALSE} to determine whether
partial matching is allowed in \code{grade_this_code()}. Sets the
\code{gradethis.allow_partial_matching} option.}

\item{exercise.checker}{Function used to check exercise answers
(e.g., \code{gradethis::grade_learnr()}).}

\item{exercise.timelimit}{Number of seconds to limit execution time to
(defaults to \code{30}).}

\item{compare_timelimit}{\code{\link[=pass_if_equal]{pass_if_equal()}} and \code{\link[=fail_if_equal]{fail_if_equal()}} call
\code{\link[waldo:compare]{waldo::compare()}} internally.
This helps ensure an accurate comparison, but sometimes takes a long time.
\code{compare_timelimit} is the time limit in seconds
for the execution of \code{\link[waldo:compare]{waldo::compare()}}
(defaults to 80\% of \code{exercise.timelimit}).
If the time limit is exceeded,
\code{identical()} is used instead of \code{\link[waldo:compare]{waldo::compare()}}.}

\item{exercise.error.check.code}{A string containing R code to use for checking
code when an exercise evaluation error occurs (e.g., \code{"gradethis::grade_code()"}).}

\item{fail_code_feedback}{Deprecated. Use \code{maybe_code_feedback}.}
}
\value{
Invisibly returns the global options as they were prior to setting
them with \code{gradethis_setup()}.
}
\description{
To use \pkg{gradethis} in your \pkg{learnr} tutorial, you only need to call
\code{library(gradethis)} in your tutorial's setup chunk.

\if{html}{\out{<div class="sourceCode">}}\preformatted{```\{r setup\}
library(learnr)
library(gradethis)
```
}\if{html}{\out{</div>}}

Use \code{gradethis_setup()} to change the default options suggested by gradethis.
This function also describes in detail each of the global options available
for customization in the gradethis package. Note that you most likely do not
want to change the defaults values for the learnr tutorial options that are
prefixed with \code{exercise.}. Each of the gradethis-specific arguments sets a
global option with the same name, prefixed with \code{gradethis.}. For example,
\code{pass} sets \code{gradethis.pass}.
}
\section{Global package options}{


These global package options can be set by \code{\link[=gradethis_setup]{gradethis_setup()}} or by
directly setting the global option. The default values set for each option
when gradethis is loaded are shown below.\tabular{ll}{
   Option \tab Default Value \cr
   \code{gradethis.pass} \tab \code{"{gradethis::random_praise()} Correct!"} \cr
   \code{gradethis.pass.praise} \tab \code{FALSE} \cr
   \code{gradethis.fail} \tab \code{"Incorrect.{gradethis::maybe_code_feedback()} {gradethis::random_encouragement()}"} \cr
   \code{gradethis.fail.hint} \tab \code{FALSE} \cr
   \code{gradethis.fail.encourage} \tab \code{FALSE} \cr
   \code{gradethis.maybe_code_feedback} \tab \code{TRUE} \cr
   \code{gradethis.maybe_code_feedback.before} \tab \code{" "} \cr
   \code{gradethis.maybe_code_feedback.after} \tab \code{NULL} \cr
   \code{gradethis.code_correct} \tab \code{NULL} \cr
   \code{gradethis.code_incorrect} \tab \code{"{gradethis::pipe_warning()}{gradethis::code_feedback()} {gradethis::random_encouragement()}"} \cr
   \code{gradethis.pipe_warning} \tab \code{"I see that you are using pipe operators (e.g. \%>\%), so I want to let you know that this is how I am interpreting your code before I check it:\\n\\n```r\\n{.user_code_unpiped}\\n```\\n\\n"} \cr
   \code{gradethis.grading_problem.message} \tab \code{"A problem occurred with the grading code for this exercise."} \cr
   \code{gradethis.grading_problem.type} \tab \code{"warning"} \cr
   \code{gradethis.allow_partial_matching} \tab \code{NULL} \cr
   \code{gradethis.error_checker.message} \tab \code{"An error occurred with your code:\\n\\n```\\n{.error$message}\\n```\\n\\n\\n"} \cr
   \code{gradethis.compare_timelimit} \tab \code{NULL} \cr
}
}

\examples{
# Not run in package documentation because this function changes global opts
if (FALSE) {
  old_opts <- gradethis_setup(
    pass = "Great work!",
    fail = "{random_encouragement()}"
  )
}

# Use getOption() to see the default value
getOption("gradethis.pass")
getOption("gradethis.maybe_code_feedback")
}
\seealso{
\code{\link[=gradethis_exercise_checker]{gradethis_exercise_checker()}}
}
