% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading_demo.R
\name{grading_demo}
\alias{grading_demo}
\title{Grading Demo}
\usage{
grading_demo()
}
\description{
If you are using the RStudio IDE, \code{grading_demo()} opens an example learnr
file that demonstrates how to use the grader package to check student code.
}
\details{
The tutorial sets the learnr exercise. checker option to
\code{grade_learnr()} in the document's setup chunk.
It then uses three different exercise checking methods: \code{\link{check_result}}, \code{\link{test_result}}, and \code{\link{check_code}}.
To use a checking method, follow the exercise chunk with a chunk whose label
matches the label of the exercise chunk (ex: \code{myexercise}) but includes the suffix
\code{-check} (ex: \code{myexercise-check}). Call any checking method in that chunk. To ensure that
checking method can provide informative feedback, you may provide custom \code{correct} and \code{incorrect} messages.

If you are not using RStudio IDE, you can access the demo file at \code{system.file("extdata", "grading-demo/grading-demo.Rmd", package = "grader")}.
}
