% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass_fail_if.R
\name{pass_if}
\alias{pass_if}
\alias{fail_if}
\title{Condition definition}
\usage{
pass_if(x, message = NULL)

fail_if(x, message = NULL)
}
\arguments{
\item{x}{A formula, function, or value, that returns \code{TRUE} or \code{FALSE}.
When comparing objects that are greater than length 1 (e.g., vectors, dataframes, matricies, etc)
A boolean vector will be returned if the user uses \code{==}, not a single boolean value.
\code{grader} will run the vector through \code{all(..., na.rm = TRUE)} to check for the boolean value.
It is advised that the user use \code{identical()} instead of \code{==} in this case.}

\item{message}{chracter string for message returned}
}
\description{
Wrapper around \code{\link{condition()}} that sets the proper \code{correct} argument.
This allows the instructor to specify just the comparison code, \code{x},
and \code{message}.
}
\section{Functions}{
\itemize{
\item \code{pass_if}: a condition that if matched means the student provided result is correct

\item \code{fail_if}: a condition that if matched means the student provided result is incorrect
}}

