% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{pipe_warning}
\alias{pipe_warning}
\title{Pipe Warning Message}
\usage{
pipe_warning(message = getOption("gradethis.pipe_warning"), .user_code = NULL)
}
\arguments{
\item{message}{A glue string containing the message}

\item{.user_code}{The user's submitted code, found in \code{env} if \code{NULL}}
}
\description{
Creates a warning message when user code contains the \verb{\%>\%}.
}
