% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Random praise and encouragement}
\description{
Generate a random praise or encouragement phrase.
This can be used in conjunction with
\code{\link[glue:glue]{glue::glue()}} to generate praise or encouragement within feedback to users.
}
\examples{
replicate(5, glue::glue("Random praise: {random_praise()}"))
replicate(5, glue::glue("Random encouragement: {random_encouragement()}"))

}
\seealso{
\code{\link[learnr:random_praise]{learnr::random_praise()}}, \code{\link[learnr:random_praise]{learnr::random_encouragement()}}
}
