% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{random_praise}
\alias{random_praise}
\alias{random_encourage}
\title{Random praise and encouragement}
\usage{
random_praise()

random_encourage()
}
\description{
Generate a random praise or encouragement phrase.
This can be used in conjunction with
\verb{glue::[glue][glue::glue()]} to generate praise or encouragement within feedback to users.
}
\section{Functions}{
\itemize{
\item \code{random_praise}: Returns a random praise message

\item \code{random_encourage}: Returns a random encouragement message
}}

\examples{
replicate(5, glue::glue("Random praise: {random_praise()}"))
replicate(5, glue::glue("Random encouragement: {random_encourage()}"))
}
