% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_result.R
\name{tests}
\alias{tests}
\alias{test}
\title{Tests to check}
\usage{
tests(...)

test(message, fn)
}
\arguments{
\item{...}{a set of functions that will accept the evaluated user solution. If the test fails, it should throw an error with the message to display.}

\item{message}{Message to report back if the test throws an error.}

\item{fn}{function to execute against the user solution.  If the test fails, it should throw an error to display the \code{message} provided.}
}
\description{
Collect a set of test to execute against a user's result value
}
\examples{

tests(
  function(your_answer) {
    checkmate::expect_function(your_answer, args = c("x"))
  },
  test(
    # use a custom error message
    "Make sure your function returns a number!",
    function(your_answer) {
      checkmate::expect_number(your_answer(1))
    }
  ),
  function(your_answer) {
    testthat::expect_equal(your_answer(0), NaN)
  },
  function(your_answer) {
    testthat::expect_equal(your_answer(1:10), sqrt(log(1:10)))
  }
)
}
